+++ src/Player/PlayerBase.cpp
@@ -1,7 +1,7 @@
 ﻿/*
  * Copyright (c) 2016 The ZLMediaKit project authors. All Rights Reserved.
  *
- * This file is part of ZLMediaKit(https://github.com/xiongziliang/ZLMediaKit).
+ * This file is part of ZLMediaKit(https://github.com/xia-chu/ZLMediaKit).
  *
  * Use of this source code is governed by MIT license that can be found in the
  * LICENSE file in the root of the source tree. All contributing project authors
@@ -64,7 +64,7 @@ PlayerBase::PlayerBase() {
 
 ///////////////////////////Demuxer//////////////////////////////
 bool Demuxer::isInited(int analysisMs) {
-    if(analysisMs && _ticker.createdTime() > analysisMs){
+    if(analysisMs && _ticker.createdTime() > (uint64_t)analysisMs){
         //analysisMs毫秒后强制初始化完毕
         return true;
     }
